# [schema] Faster schema loading via caching alongside plugins