--- a/pkg/codegen/importer/language_test.go
+++ b/pkg/codegen/importer/language_test.go
@@ -28,16 +28,11 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/config"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestGenerateLanguageDefinition(t *testing.T) {
 	t.Parallel()
 	loader := schema.NewPluginLoader(utils.NewHost(testdataPath))
-	t.Cleanup(func() {
-		err := loader.Close()
-		require.NoError(t, err)
-	})
 
 	cases, err := readTestCases("testdata/cases.json")
 	if !assert.NoError(t, err) {