--- a/pkg/codegen/schema/loader.go
+++ b/pkg/codegen/schema/loader.go
@@ -1,6 +1,7 @@
 package schema
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"io"
@@ -9,6 +10,10 @@ import (
 	"regexp"
 	"sync"
 	"time"
+
+	"github.com/edsrzf/mmap-go"
+	"github.com/hashicorp/go-multierror"
+	"github.com/natefinch/atomic"
 
 	"github.com/blang/semver"
 	"github.com/segmentio/encoding/json"
@@ -25,6 +30,7 @@ type Loader interface {
 
 type ReferenceLoader interface {
 	Loader
+	io.Closer
 
 	LoadPackageReference(pkg string, version *semver.Version) (PackageReference, error)
 }
@@ -34,6 +40,20 @@ type pluginLoader struct {
 
 	host    plugin.Host
 	entries map[string]PackageReference
+	files   []*os.File
+	mmaps   []mmap.MMap
+
+	cacheOptions pluginLoaderCacheOptions
+}
+
+// Caching options intended for benchmarking or debugging:
+type pluginLoaderCacheOptions struct {
+	// useEntriesCache enables in-memory re-use of packages
+	disableEntryCache bool
+	// useFileCache enables skipping plugin loading when possible and caching JSON schemas to files.
+	disableFileCache bool
+	// useMmap enables the use of memory mapped IO to avoid copying the JSON schema when using
+	disableMmap bool
 }
 
 func NewPluginLoader(host plugin.Host) ReferenceLoader {
@@ -43,12 +63,35 @@ func NewPluginLoader(host plugin.Host) ReferenceLoader {
 	}
 }
 
+func newPluginLoaderWithOptions(host plugin.Host, cacheOptions pluginLoaderCacheOptions) ReferenceLoader {
+	return &pluginLoader{
+		host:    host,
+		entries: map[string]PackageReference{},
+
+		cacheOptions: cacheOptions,
+	}
+}
+
 func (l *pluginLoader) getPackage(key string) (PackageReference, bool) {
-	l.m.RLock()
-	defer l.m.RUnlock()
+	if l.cacheOptions.disableEntryCache {
+		return nil, false
+	}
 
 	p, ok := l.entries[key]
 	return p, ok
+}
+
+func (l *pluginLoader) setPackage(key string, p PackageReference) PackageReference {
+	if l.cacheOptions.disableEntryCache {
+		return p
+	}
+
+	if p, ok := l.entries[key]; ok {
+		return p
+	}
+
+	l.entries[key] = p
+	return p
 }
 
 // ensurePlugin downloads and installs the specified plugin if it does not already exist.
@@ -134,6 +177,27 @@ func (l *pluginLoader) ensurePlugin(pkg string, version *semver.Version) error {
 	return nil
 }
 
+func (l *pluginLoader) Close() error {
+	var result error
+	for _, m := range l.mmaps {
+		if err := m.Unmap(); err != nil {
+			result = multierror.Append(result, err)
+		}
+	}
+
+	for _, f := range l.files {
+		if err := f.Close(); err != nil {
+			result = multierror.Append(result, err)
+		}
+	}
+
+	if err := l.host.Close(); err != nil {
+		result = multierror.Append(result, err)
+	}
+
+	return result
+}
+
 func (l *pluginLoader) LoadPackage(pkg string, version *semver.Version) (*Package, error) {
 	ref, err := l.LoadPackageReference(pkg, version)
 	if err != nil {
@@ -159,24 +223,15 @@ func schemaIsEmpty(schemaBytes []byte) bool {
 }
 
 func (l *pluginLoader) LoadPackageReference(pkg string, version *semver.Version) (PackageReference, error) {
-	key := packageIdentity(pkg, version)
+	l.m.Lock()
+	defer l.m.Unlock()
 
+	key := packageIdentity(pkg, version)
 	if p, ok := l.getPackage(key); ok {
 		return p, nil
 	}
 
-	if err := l.ensurePlugin(pkg, version); err != nil {
-		return nil, err
-	}
-
-	provider, err := l.host.Provider(tokens.Package(pkg), version)
-	if err != nil {
-		return nil, err
-	}
-	contract.Assert(provider != nil)
-
-	schemaFormatVersion := 0
-	schemaBytes, err := provider.GetSchema(schemaFormatVersion)
+	schemaBytes, version, err := l.loadSchemaBytes(pkg, version)
 	if err != nil {
 		return nil, err
 	}
@@ -185,21 +240,13 @@ func (l *pluginLoader) LoadPackageReference(pkg string, version *semver.Version)
 	}
 
 	var spec PartialPackageSpec
-	if _, err = json.Parse(schemaBytes, &spec, json.ZeroCopy); err != nil {
+	if _, err := json.Parse(schemaBytes, &spec, json.ZeroCopy); err != nil {
 		return nil, err
 	}
 
 	// Insert a version into the spec if the package does not provide one
-	if spec.PackageInfoSpec.Version == "" {
-		if version == nil {
-			providerInfo, err := provider.GetPluginInfo()
-			if err == nil {
-				version = providerInfo.Version
-			}
-		}
-		if version != nil {
-			spec.PackageInfoSpec.Version = version.String()
-		}
+	if version != nil && spec.PackageInfoSpec.Version == "" {
+		spec.PackageInfoSpec.Version = version.String()
 	}
 
 	p, err := importPartialSpec(spec, nil, l)
@@ -207,15 +254,7 @@ func (l *pluginLoader) LoadPackageReference(pkg string, version *semver.Version)
 		return nil, err
 	}
 
-	l.m.Lock()
-	defer l.m.Unlock()
-
-	if p, ok := l.entries[pkg]; ok {
-		return p, nil
-	}
-	l.entries[key] = p
-
-	return p, nil
+	return l.setPackage(key, p), nil
 }
 
 func LoadPackageReference(loader Loader, pkg string, version *semver.Version) (PackageReference, error) {
@@ -227,4 +266,111 @@ func LoadPackageReference(loader Loader, pkg string, version *semver.Version) (P
 		return nil, err
 	}
 	return p.Reference(), nil
+}
+
+func (l *pluginLoader) loadSchemaBytes(pkg string, version *semver.Version) ([]byte, *semver.Version, error) {
+	if err := l.ensurePlugin(pkg, version); err != nil {
+		return nil, nil, err
+	}
+
+	pluginInfo, err := l.host.ResolvePlugin(workspace.ResourcePlugin, pkg, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	if version == nil {
+		version = pluginInfo.Version
+	}
+
+	if pluginInfo.SchemaPath != "" {
+		schemaBytes, ok := l.loadCachedSchemaBytes(pkg, pluginInfo.SchemaPath, pluginInfo.SchemaTime)
+		if ok {
+			return schemaBytes, nil, nil
+		}
+	}
+
+	schemaBytes, provider, err := l.loadPluginSchemaBytes(pkg, version)
+	if err != nil {
+		return nil, nil, fmt.Errorf("Error loading schema from plugin: %w", err)
+	}
+
+	if pluginInfo.SchemaPath != "" {
+		err = atomic.WriteFile(pluginInfo.SchemaPath, bytes.NewReader(schemaBytes))
+		if err != nil {
+			return nil, nil, fmt.Errorf("Error writing schema from plugin to cache: %w", err)
+		}
+	}
+
+	if version == nil {
+		info, err := provider.GetPluginInfo()
+		if err != nil {
+			// Nonfatal
+		}
+		version = info.Version
+	}
+
+	return schemaBytes, version, nil
+}
+
+func (l *pluginLoader) loadPluginSchemaBytes(pkg string, version *semver.Version) ([]byte, plugin.Provider, error) {
+	provider, err := l.host.Provider(tokens.Package(pkg), version)
+	if err != nil {
+		return nil, nil, err
+	}
+	contract.Assert(provider != nil)
+
+	schemaFormatVersion := 0
+	schemaBytes, err := provider.GetSchema(schemaFormatVersion)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return schemaBytes, provider, nil
+}
+
+func (l *pluginLoader) loadCachedSchemaBytes(pkg string, path string, schemaTime time.Time) ([]byte, bool) {
+	if l.cacheOptions.disableFileCache {
+		return nil, false
+	}
+
+	success := false
+	schemaFile, err := os.OpenFile(path, os.O_RDONLY, 0644)
+	defer func() {
+		if success {
+			l.files = append(l.files, schemaFile)
+		} else {
+			schemaFile.Close()
+		}
+	}()
+	if err != nil {
+		return nil, success
+	}
+
+	stat, err := schemaFile.Stat()
+	if err != nil {
+		return nil, success
+	}
+	cachedAt := stat.ModTime()
+
+	if schemaTime.After(cachedAt) {
+		return nil, success
+	}
+
+	if l.cacheOptions.disableMmap {
+		data, err := io.ReadAll(schemaFile)
+		if err != nil {
+			return nil, success
+		}
+		success = true
+		return data, success
+	}
+
+	schemaMmap, err := mmap.Map(schemaFile, mmap.RDONLY, 0)
+	if err != nil {
+		return nil, success
+	}
+	success = true
+	l.mmaps = append(l.mmaps, schemaMmap)
+
+	return schemaMmap, success
 }