--- a/pkg/codegen/schema/loader_schema_test.go
+++ b/pkg/codegen/schema/loader_schema_test.go
@@ -8,6 +8,7 @@ import (
 
 func TestEmptySchemaResponse(t *testing.T) {
 	t.Parallel()
+
 	assert.True(t, schemaIsEmpty([]byte("{}")))
 	assert.True(t, schemaIsEmpty([]byte("{  }")))
 	assert.True(t, schemaIsEmpty([]byte("{	}")))           // tab character