--- a/pkg/resource/deploy/deploytest/pluginhost.go
+++ b/pkg/resource/deploy/deploytest/pluginhost.go
@@ -19,6 +19,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"path/filepath"
 	"sync"
 
 	"github.com/blang/semver"
@@ -57,12 +58,19 @@ func WithGrpc(p *PluginLoader) {
 	p.useGRPC = true
 }
 
+func WithPath(path string) func(p *PluginLoader) {
+	return func(p *PluginLoader) {
+		p.path = path
+	}
+}
+
 type PluginLoader struct {
 	kind         workspace.PluginKind
 	name         string
 	version      semver.Version
 	load         LoadPluginFunc
 	loadWithHost LoadPluginWithHostFunc
+	path         string
 	useGRPC      bool
 }
 
@@ -387,6 +395,25 @@ func (host *pluginHost) CloseProvider(provider plugin.Provider) error {
 }
 func (host *pluginHost) EnsurePlugins(plugins []workspace.PluginInfo, kinds plugin.Flags) error {
 	return nil
+}
+func (host *pluginHost) ResolvePlugin(
+	kind workspace.PluginKind, name string, version *semver.Version) (*workspace.PluginInfo, error) {
+	for _, v := range host.pluginLoaders {
+		if v.kind == kind && v.name == name {
+			// TODO: for multi-version testing, support multiple versions of plugins concurrently. Not
+			// allowed at present time.
+			return &workspace.PluginInfo{
+				Kind:       kind,
+				Name:       name,
+				Path:       v.path,
+				Version:    &v.version,
+				SchemaPath: filepath.Join(v.path, name+".json"),
+				// SchemaTime not set as caching is indefinite.
+			}, nil
+		}
+	}
+
+	return nil, nil
 }
 func (host *pluginHost) GetRequiredPlugins(info plugin.ProgInfo,
 	kinds plugin.Flags) ([]workspace.PluginInfo, error) {