--- a/sdk/go/common/workspace/plugins.go
+++ b/sdk/go/common/workspace/plugins.go
@@ -485,6 +485,8 @@ type PluginInfo struct {
 	LastUsedTime      time.Time       // the last time the plugin was used.
 	PluginDownloadURL string          // an optional server to use when downloading this plugin.
 	PluginDir         string          // if set, will be used as the root plugin dir instead of ~/.pulumi/plugins.
+	SchemaPath        string          // if set, used as the path for loading and caching the schema
+	SchemaTime        time.Time       // if set and newer than the file at SchemaPath, used to invalidate a cached schema
 }
 
 // Dir gets the expected plugin directory for this plugin.
@@ -584,20 +586,47 @@ func (info *PluginInfo) SetFileMetadata(path string) error {
 
 	// Next, get the size from the directory (or, if there is none, just the file).
 	size, err := getPluginSize(path)
-	if err != nil {
-		return errors.Wrapf(err, "getting plugin dir %s size", path)
+	if err == nil {
+		info.Size = size
+	} else {
+		logging.V(6).Infof("unable to get plugin dir size for %s: %v", path, err)
 	}
-	info.Size = size
 
 	// Next get the access times from the plugin binary itself.
 	tinfo := times.Get(file)
 
-	if tinfo.HasBirthTime() {
-		info.InstallTime = tinfo.BirthTime()
+	if tinfo.HasChangeTime() {
+		info.InstallTime = tinfo.ChangeTime()
+	} else {
+		info.InstallTime = tinfo.ModTime()
 	}
 
 	info.LastUsedTime = tinfo.AccessTime()
+
+	if info.Kind == ResourcePlugin {
+		info.SetSchemaMetadata()
+	}
+
 	return nil
+}
+
+func (info *PluginInfo) SetSchemaMetadata() {
+	binpath, err := info.FilePath()
+	if err != nil {
+		return
+	}
+	bintime, err := times.Stat(binpath)
+	if err != nil {
+		return
+	}
+
+	dir, err := info.DirPath()
+	if err != nil {
+		return
+	}
+
+	info.SchemaPath = filepath.Join(dir, "schema-"+info.Name+".json")
+	info.SchemaTime = bintime.ModTime()
 }
 
 func interpolateURL(serverURL string, version semver.Version, os, arch string) string {
@@ -1015,12 +1044,13 @@ func getPlugins(dir string, skipMetadata bool) ([]PluginInfo, error) {
 	for _, file := range files {
 		// Skip anything that doesn't look like a plugin.
 		if kind, name, version, ok := tryPlugin(file); ok {
+			path := filepath.Join(dir, file.Name())
 			plugin := PluginInfo{
 				Name:    name,
 				Kind:    kind,
 				Version: &version,
+				Path:    path,
 			}
-			path := filepath.Join(dir, file.Name())
 			if _, err := os.Stat(fmt.Sprintf("%s.partial", path)); err == nil {
 				// Skip it if the partial file exists, meaning the plugin is not fully installed.
 				continue
@@ -1044,6 +1074,53 @@ func getPlugins(dir string, skipMetadata bool) ([]PluginInfo, error) {
 // using standard semver sorting rules.  A plugin may be overridden entirely by placing it on your $PATH, though it is
 // possible to opt out of this behavior by setting PULUMI_IGNORE_AMBIENT_PLUGINS to any non-empty value.
 func GetPluginPath(kind PluginKind, name string, version *semver.Version) (string, string, error) {
+	info, path, err := getPluginInfoOrPath(kind, name, version, true /* skipMetadata */)
+	if err != nil {
+		return "", "", err
+	}
+
+	if info != nil {
+		matchDir, err := info.DirPath()
+		if err != nil {
+			return "", "", err
+		}
+
+		matchPath, err := info.FilePath()
+		if err != nil {
+			return "", "", err
+		}
+
+		return matchDir, matchPath, nil
+	}
+
+	return "", path, err
+}
+
+func GetPluginInfo(kind PluginKind, name string, version *semver.Version) (*PluginInfo, error) {
+	info, path, err := getPluginInfoOrPath(kind, name, version, false)
+	if err != nil {
+		return nil, err
+	}
+
+	if info != nil {
+		return info, nil
+	}
+
+	info = &PluginInfo{
+		Kind: kind,
+		Name: name,
+		Path: filepath.Dir(path),
+	}
+
+	return info, nil
+}
+
+// getPluginInfoOrPath searches for a compatible plugin kind, name, and version and returns either:
+//  * if found as an ambient plugin, nil and the path to the executable
+//  * if found in the pulumi dir's installed plugins, a PluginInfo and path to the executable
+//  * an error in all other cases.
+func getPluginInfoOrPath(
+	kind PluginKind, name string, version *semver.Version, skipMetadata bool) (*PluginInfo, string, error) {
 	var filename string
 
 	// We currently bundle some plugins with "pulumi" and thus expect them to be next to the pulumi binary. We
@@ -1061,7 +1138,7 @@ func GetPluginPath(kind PluginKind, name string, version *semver.Version) (strin
 		filename = (&PluginInfo{Kind: kind, Name: name, Version: version}).FilePrefix()
 		if path, err := exec.LookPath(filename); err == nil {
 			logging.V(6).Infof("GetPluginPath(%s, %s, %v): found on $PATH %s", kind, name, version, path)
-			return "", path, nil
+			return nil, path, nil
 		}
 	}
 
@@ -1086,7 +1163,7 @@ func GetPluginPath(kind PluginKind, name string, version *semver.Version) (strin
 						logging.V(6).Infof("GetPluginPath(%s, %s, %v): found next to current executable %s",
 							kind, name, version, candidate)
 
-						return "", candidate, nil
+						return nil, candidate, nil
 					}
 				}
 			}
@@ -1094,9 +1171,15 @@ func GetPluginPath(kind PluginKind, name string, version *semver.Version) (strin
 	}
 
 	// Otherwise, check the plugin cache.
-	plugins, err := GetPlugins()
+	var plugins []PluginInfo
+	var err error
+	if skipMetadata {
+		plugins, err = GetPlugins()
+	} else {
+		plugins, err = GetPluginsWithMetadata()
+	}
 	if err != nil {
-		return "", "", errors.Wrapf(err, "loading plugin list")
+		return nil, "", errors.Wrapf(err, "loading plugin list")
 	}
 
 	var match *PluginInfo
@@ -1104,7 +1187,7 @@ func GetPluginPath(kind PluginKind, name string, version *semver.Version) (strin
 		logging.V(6).Infof("GetPluginPath(%s, %s, %s): enabling new plugin behavior", kind, name, version)
 		candidate, err := SelectCompatiblePlugin(plugins, kind, name, semver.MustParseRange(version.String()))
 		if err != nil {
-			return "", "", NewMissingError(PluginInfo{
+			return nil, "", NewMissingError(PluginInfo{
 				Name:    name,
 				Kind:    kind,
 				Version: version,
@@ -1139,20 +1222,16 @@ func GetPluginPath(kind PluginKind, name string, version *semver.Version) (strin
 	}
 
 	if match != nil {
-		matchDir, err := match.DirPath()
-		if err != nil {
-			return "", "", err
-		}
 		matchPath, err := match.FilePath()
 		if err != nil {
-			return "", "", err
+			return nil, "", err
 		}
 
 		logging.V(6).Infof("GetPluginPath(%s, %s, %v): found in cache at %s", kind, name, version, matchPath)
-		return matchDir, matchPath, nil
+		return match, matchPath, nil
 	}
 
-	return "", "", NewMissingError(PluginInfo{
+	return nil, "", NewMissingError(PluginInfo{
 		Name:    name,
 		Kind:    kind,
 		Version: version,