… CHANGELOG_PENDING.md
… pkg/codegen/testing/test/sdk_driver.go
… pkg/codegen/testing/test/testdata/eks.json
… pkg/codegen/testing/test/testdata/external-go-import-aliases/schema.json
… pkg/codegen/testing/test/testdata/external-node-compatibility/schema.json
… pkg/codegen/testing/test/testdata/external-resource-schema/schema.json
… pkg/codegen/testing/test/testdata/plain-schema-gh6957/schema.json
… pkg/go.sum
… pkg/resource/deploy/providers/registry_test.go
… tests/go.mod
… tests/go.sum