Implements caching of schemas in package loader. Package loader searches for plugins & before calling GetSchema, attempts to read a `schema-${pkg}.json` file. On cache miss or when the plugin is newer than the cache, the file is updated via GetSchema call.

Tested against locally installed packages as well as plugin cache located packages, e.g.: when `pulumi-resource-azure-native` is in $GOBIN, this caches the schema to `$GOBIN/schema-azure-native.json`.

A micro benchmark shows substantial improvements in wall time and bytes of memory allocated/copied, and a combined benchmark of this PR along with PRs to Azure Native & YAML reduced the time for a `pulumi preview` on an Azure Native stack down to 2.5 seconds from 12, and down to 340MiB peak memory usage from ~2GiB.

A microbenchmark shows the reduction in `loader_test.go`:

```
enchmarkLoadPackageReference/no-cache-8      1   2231342910 ns/op  351691352 B/op   7063 allocs/op
BenchmarkLoadPackageReference/file-cache-8   4    281257858 ns/op  631195948 B/op   2998 allocs/op
BenchmarkLoadPackageReference/mmap-cache-8   7    157173553 ns/op   15967628 B/op   2934 allocs/op
```
no cache: starting a plugin process and calling GetSchema() on it in every iteration. the memory usage is misleading as a result - it can't count the subprocess.

file cache: using a schema.json file colocated with the plugin. (this benchmark is pre-warmed, but it would be created if this cache misses)

mmap cache: instead of using io.ReadAll (copying) bytes, use mmap with O_RDONLY

Related PRs:
- https://github.com/pulumi/pulumi/pull/9620
- https://github.com/pulumi/pulumi-azure-native/pull/1689/
- https://github.com/pulumi/pulumi-yaml/pull/230