@iwahbe 2022-05-26
> @iwahbe I ran into a really strange issue with fixing the binder on this one - we have some unit tests that reference `google-native` types via $ref. Not just that, but there's a few different versions of the plugin in the json pointers. Somehow these tests pass on a clean system without these changes, but fail after. Installing the google native resource plugin causes tests to pass, but I also don't see where that could be occurring.
> 
> The strange part is that I can't find where in our code we're getting the google native schema _or_ where there's a branch that's allowing the $ref to fail _or_ where the plugin is installed.

That sounds super weird. Is it possible that we are doing something with pointer equality: attempt to load several versions of the google-native, fail and assign everything to nil pointers. With changes results in failure because you can't load the package. That would be my best guess, but I would need to see where the errors are to be more helpful.