@AaronFriel 2022-05-26
@pgavlin Without schema caching, but with your PR to YAML here https://github.com/pulumi/pulumi-yaml/pull/230

755 MiB peak resident memory
2.95s wall time

Re-running the experiment after checking this branch out but with an empty plugin cache (all I've done is `pulumi plugin install ...`):

734 MiB 
3.05s

And second run after the schema is cached:

342 MiB peak
2.13s