@AaronFriel 2022-05-26
Timing and memory usage is very consistent, over 3 runs without this PR, I saw 755, 754, 758 MiB. Wall times are less consistent, 2.9-3.7s.

With this PR, over 3 runs: 340, 334, 342 MiB and a range of 2.170-2.191s.