@AaronFriel 2022-05-26
Building YAML & pulumi with that branch:

| RSS (MiB) | time (s) |
|----------:|---------:|
|       751 |    2.911 |
|       738 |    2.927 |
|       734 |    3.064 |