@AaronFriel 2022-05-26
Here's process tree memory usage for the 774 MiB run:

```
🌊 gotta-watch-em-all: Reached a new high water mark of 774 MiB, 37 MiB greater than before!
🌊 process                                                                          private MiB    total MiB
🌊 pulumi (22359)                                                                         61MiB       774MiB
🌊   pulumi-language-yaml (22380)                                                        250MiB       704MiB
🌊     pulumi-resource-azure-native (22392)                                              424MiB       424MiB
🌊     pulumi-resource-docker (22417)                                                     28MiB        28MiB
🌊   pulumi-resource-azure-native (22428)                                                  9MiB         9MiB
```

The bulk of the usage is in the azure native provider - which is also built with a zero-copy GetSchema using the embedded string:
- https://github.com/pulumi/pulumi-azure-native/pull/1689/