@AaronFriel 2022-05-26
> It feels somewhat more reliable to me to have plugins distribute the schema file next to the binary (though that does have its own set of issues).

Let's chat about this internally, I agree that a change like this has ramifications outside the engine & providers need to be included in that.

> Maybe it would be enough to have the plugin installer remove any cached schema during installation, but I'm not sure.

We wouldn't need to, but with the above, we need to agree on a naming convention for the file:

```
--- /home/friel/.pulumi/plugins/resource-aws-v4.36.0 ---------
                         /..
  343.3 MiB [##########]  pulumi-resource-aws
   15.8 MiB [          ]  schema-aws.json                            
   44.0 KiB [          ]  CHANGELOG.md
   12.0 KiB [          ]  LICENSE
   12.0 KiB [          ]  README.md
```