@AaronFriel 2022-06-02
@pgavlin this is ready for review now, I've pared down the caching to only work on plugins installed in the plugin dir (and deploytest) and pushed decisions about where and when to invalidate the cache out to the workspace / plugin host. Whatever the plugin host returns from `ResolvePlugin` is authoritative.